Doc: Update API Desc to OS
Problem: As the project was transformed into an opensource project the
API description should also be in english. Translating the
documentation a few descriptions were changed to make the documentation
easier to understand.

BREAKING CHANGES: The new english version completly replaces the old
german version with a few changes in the description.